% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{cred}
\alias{cred}
\title{Title}
\usage{
cred(
  PIP,
  coverage = 0.9,
  purity = 0.5,
  LD = NULL,
  V = NULL,
  Dsq = NULL,
  n = NULL,
  dedup = TRUE
)
}
\arguments{
\item{PIP}{p x L PIP matrix output by susieK}

\item{coverage}{desired coverage level for each credible set}

\item{purity}{sets with minimum absolute correlation < purity are removed}

\item{LD}{LD matrix (equal to X'X/n)}

\item{V}{precomputed p x p matrix of eigenvectors of X'X}

\item{Dsq}{precomputed length-p vector of eigenvalues of X'X}

\item{n}{sample size}

\item{dedup}{remove duplicate CS's}
}
\description{
Title
}
