% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spa_helpers.R
\name{d2_wcgf}
\alias{d2_wcgf}
\title{\code{d2_wcgf} is a function computing the hessian of the weighted cumulant
generating function (WCGF) of distributions from GLM family}
\usage{
d2_wcgf(s, P, W, fam)
}
\arguments{
\item{s}{The point where the CGF will be computed.}

\item{P}{A vector containing the parameter values of the family of distributions.}

\item{W}{A vector containing the weights.}

\item{fam}{The GLM family which includes the distribution whose CGF is being
evaluated (values can be \code{gaussian}, \code{binomial}, \code{poisson}, etc).}
}
\value{
Second derivative of CGF of the weighted distribution evaluated at \code{s}.
}
\description{
\code{d2_wcgf} is a function computing the hessian of the weighted cumulant
generating function (WCGF) of distributions from GLM family
}
\keyword{internal}
