% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_X.R
\name{fit_X_shrinkage}
\alias{fit_X_shrinkage}
\title{Fit L(X) using knockoffs default
Code adapted from knockoff::create.second_order()}
\usage{
fit_X_shrinkage(data, shrink = T)
}
\arguments{
\item{data}{An object contaning a design matrix X.}

\item{shrink}{shrink or not}
}
\value{
An object with fields \code{generate_resamples} and \code{conditional_mean}
}
\description{
Fit L(X) using knockoffs default
Code adapted from knockoff::create.second_order()
}
