% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_y_given_X.R
\name{fit_y_given_X_lasso_binomial}
\alias{fit_y_given_X_lasso_binomial}
\title{Fit y given X via lasso}
\usage{
fit_y_given_X_lasso_binomial(data)
}
\arguments{
\item{data}{A list containing a matrix X and vector y.}
}
\value{
A fitted E(y|X) function
}
\description{
Fit y given X via lasso
}
