% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_y_given_X.R
\name{fit_y_given_X_post_lasso}
\alias{fit_y_given_X_post_lasso}
\title{Fit y given X via post-lasso}
\usage{
fit_y_given_X_post_lasso(data)
}
\arguments{
\item{data}{A list containing a matrix X and vector y.}
}
\value{
A fitted E(y|X) function
}
\description{
Fit y given X via post-lasso
}
