% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{generate_data}
\alias{generate_data}
\title{Title}
\usage{
generate_data(
  n,
  p,
  s,
  joint_X,
  y_given_X,
  X_hyperparams,
  y_given_X_hyperparams,
  amplitude,
  ground_truth
)
}
\arguments{
\item{n}{}

\item{p}{}

\item{s}{}

\item{joint_X}{}

\item{y_given_X}{}

\item{X_hyperparams}{}

\item{y_given_X_hyperparams}{}

\item{amplitude}{}

\item{ground_truth}{}
}
\value{
a list that includes an X matrix, y vector, and any other relevant objects (such as a tree) or information
}
\description{
Title
}
