% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{generate_data_treeWAS}
\alias{generate_data_treeWAS}
\title{Title}
\usage{
generate_data_treeWAS(n.ind, n.snps, n.snps.assoc, assoc.prob)
}
\arguments{
\item{n.ind}{}

\item{n.snps}{}

\item{n.snps.assoc}{}

\item{assoc.prob}{}
}
\description{
Title
}
