% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{generate_data_treeWAS_mod}
\alias{generate_data_treeWAS_mod}
\title{Title}
\usage{
generate_data_treeWAS_mod(
  n.ind,
  n.snps,
  n.snps.assoc,
  assoc.prob = 95,
  set = 1,
  ground_truth
)
}
\arguments{
\item{n.ind}{}

\item{n.snps}{}

\item{n.snps.assoc}{}

\item{assoc.prob}{}

\item{set}{}

\item{ground_truth}{}
}
\description{
Title
}
