% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{gmmat_binary}
\alias{gmmat_binary}
\title{Title}
\usage{
gmmat_binary(
  data,
  covar = NULL,
  phylo = FALSE,
  jaccard = FALSE,
  hamming = FALSE,
  normalized = FALSE,
  tmpdir = "gmmat_temp",
  output = "gmmat_out.txt"
)
}
\arguments{
\item{data}{a list object that contains an X matrix, y vector, phylogenetic tree (optional).
The rownames of X and names of y should align, and be a subset of tree$tip.label if tree is supplied.}

\item{covar}{covariate matrix.}

\item{phylo}{whether to use tree similarity/distance (tree must be provided).}

\item{jaccard}{whether to use jaccard similarity/distance.}

\item{hamming}{whether to use hamming similarity/distance.}

\item{normalized}{whether to use a normalized XX^T similarity/distance.}

\item{tmpdir}{path to temp directory to store output}

\item{output}{name of output file.}
}
\value{
GMMAT results output
}
\description{
Title
}
