% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{oracle_treeWAS_wrapper}
\alias{oracle_treeWAS_wrapper}
\title{Title}
\usage{
oracle_treeWAS_wrapper(
  data,
  method = c("terminal", "simultaneous", "subsequent"),
  n.snps.sim = 10000
)
}
\arguments{
\item{data}{a list object that contains an X matrix, y vector, phylogenetic tree,
and a number of mutations vector n.mts.}

\item{method}{vector of scores to compute in treeWAS}

\item{n.snps.sim}{number of samples to use}
}
\value{
output from treeWAS
}
\description{
Title
}
