% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{process_treeWAS}
\alias{process_treeWAS}
\title{Title}
\usage{
process_treeWAS(treeWAS_object)
}
\arguments{
\item{treeWAS_object}{a treeWAS output object}
}
\value{
list object with each scores p-values
}
\description{
Title
}
