% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{pyseer_wrapper}
\alias{pyseer_wrapper}
\title{Title}
\usage{
pyseer_wrapper(
  data,
  method,
  tmpdir = "pyseer_temp",
  output = "pyseer_out.txt",
  conda_bin,
  pyseer_env,
  max_dimensions = 10,
  phylo = FALSE,
  jaccard = FALSE,
  hamming = FALSE,
  normalized = FALSE,
  extra_args = NULL,
  keep_files = FALSE
)
}
\arguments{
\item{data}{a list object that contains an X matrix, y vector, phylogenetic tree (optional).
The rownames of X and names of y should align, and be a subset of tree$tip.label if tree is supplied.}

\item{method}{"fixed" or "mixed", no default.}

\item{tmpdir}{a path to directory to put temporary files.}

\item{output}{name of output file.}

\item{conda_bin}{path to the conda executable}

\item{pyseer_env}{name of conda environment}

\item{max_dimensions}{if using method = "fixed", how many MDS dimensions to use.}

\item{phylo}{whether to use tree similarity/distance (tree must be provided).}

\item{jaccard}{whether to use jaccard similarity/distance.}

\item{hamming}{whether to use hamming similarity/distance.}

\item{normalized}{whether to use a normalized XX^T similarity/distance.}

\item{extra_args}{}

\item{keep_files}{whether to keep the temp files, default FALSE.}
}
\value{
pyseer results output
}
\description{
Title
}
