% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{raw_treeWAS}
\alias{raw_treeWAS}
\title{Title}
\usage{
raw_treeWAS(treeWAS_object)
}
\arguments{
\item{treeWAS_object}{a treeWAS output object}
}
\value{
list object with each scores p-values directly from treeWAS
}
\description{
Title
}
