% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemap.R
\name{run_susie_clustered}
\alias{run_susie_clustered}
\title{Wrapper to run locus definition + SuSiE and compile into a standard output}
\usage{
run_susie_clustered(X, y, define_fn, susie_fn, var_names = NULL, ...)
}
\arguments{
\item{X}{n x p genotype matrix}

\item{y}{length-n phenotype vector}

\item{define_fn}{locus definition function, must return list with $variants and $leads}

\item{susie_fn}{susie runner function, must accept (loci, X, y, ...)}

\item{var_names}{optional length-p character vector of variant names}

\item{...}{additional arguments passed to susie_fn}
}
\value{
list with:
$pip      : length-p vector of PIPs (max across loci if overlapping)
$pip_mean : length-p vector of mean PIPs across loci
$pip_all  : length-p list of all PIP estimates per variant
$pip_n    : length-p integer vector of how many loci each variant appeared in
$cs       : list of credible sets as original variant indices
$loci     : raw loci definition
$results  : raw per-locus susie results
}
\description{
Wrapper to run locus definition + SuSiE and compile into a standard output
}
