% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemap.R
\name{run_susie_function_loci}
\alias{run_susie_function_loci}
\title{Run susie_function on each locus}
\usage{
run_susie_function_loci(
  loci,
  data,
  susie_function,
  L = 10,
  min_pip = 0.95,
  ...
)
}
\arguments{
\item{loci}{output of define_loci()}

\item{data}{list with $X (n x p), $y (length-n), $tree (phylo object or NULL)}

\item{susie_function}{choice of susie method}

\item{L}{max number of causal signals per locus (default 10)}

\item{min_pip}{PIP threshold to define a credible set (variants with PIP > threshold)}

\item{...}{additional arguments passed to susie_function}
}
\value{
list of results, one per locus, each with:
$fit         : raw susieKv2 fit object
$cs          : credible sets with original variant indices
$pip         : named vector of marginal PIPs for locus variants (original indices)
}
\description{
Run susie_function on each locus
}
