% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{scoary_wrapper}
\alias{scoary_wrapper}
\title{Title}
\usage{
scoary_wrapper(
  data,
  tmpdir = "scoary_temp",
  N = 10000,
  conda_bin,
  scoary_env,
  threads = 1,
  low.var = NULL,
  extra_args = NULL,
  keep_files = FALSE
)
}
\arguments{
\item{data}{}

\item{tmpdir}{}

\item{conda_bin}{}

\item{extra_args}{}

\item{keep_files}{}

\item{method}{}

\item{output}{}

\item{pyseer_env}{}
}
\description{
Title
}
