% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeWAS.mod.R
\name{snp.sim.Q.mod.new}
\alias{snp.sim.Q.mod.new}
\title{Aternative SNPs simulation fn.}
\usage{
snp.sim.Q.mod.new(
  n.snps = 10,
  tree = coalescent.tree.sim(100, seed = 1),
  phen.reconstruction,
  s = NULL,
  af = 10,
  snp.root = NULL,
  heatmap = FALSE,
  reconstruct = FALSE,
  dist.dna.model = "JC69",
  grp.min = 0.25,
  row.names = NULL,
  ground.truth = NULL,
  seed = 1
)
}
\arguments{
\item{n.snps}{An integer specifying the number of snps columns to be simulated.
At present, all snps are simulated to be correlated to the
provided \code{phen.reconstruction} along the provided \code{tree}.}

\item{tree}{A \code{phylo} object containing the phylogenetic tree; or, a character string,
one of \code{"NJ"}, \code{"BIONJ"} (the default), or \code{"parsimony"};
or, if NAs are present in the distance matrix, one of: \code{"NJ*"} or \code{"BIONJ*"},
specifying the method of phylogenetic reconstruction.}

\item{phen.reconstruction}{Either a character string specifying \code{"parsimony"} (the default) or \code{"ML"} (maximum likelihood)
for the ancestral state reconstruction of the phenotypic variable,
or a vector containing this reconstruction if it has been performed elsewhere.}

\item{s}{If \code{set} is 3, the \code{s} parameter controls a baseline number of substitutions to be
experienced by the phenotype and associated loci: by default, 20.}

\item{af}{If \code{set} is 3, the \code{af} parameter provides an association factor,
controlling the preference for association over non-association at associated loci:  by default, 10 (for a 10x preference).}

\item{heatmap}{A logical indicating whether to produce a heatmap of the genetic distance
between the simulated genomes of the n.ind individuals.}

\item{reconstruct}{Either a logical indicating whether to attempt to reconstruct
a phylogenetic tree using the simulated genetic data, or one of c("UPGMA", "nj", "ml")
to specify that tree reconstruction is desired by one of these three methods
(Unweighted Pair Group Method with Arithmetic Mean, Neighbour-Joining, Maximum-Likelihood).}

\item{dist.dna.model}{A character string specifying the type of model to use in reconstructing the phylogenetic tree for
calculating the genetic distance between individual genomes, only used if \code{tree} is
a character string (see ?dist.dna).}

\item{grp.min}{(Not yet (re-)implemented in this function.)
An optional number between 0.1 and 0.9 to control the proportional size of the smaller phenotypic group.}

\item{row.names}{An optional vector containing row names for the individuals to be simulated.}

\item{seed}{An optional integer to control the pseudo-randomisation process and allow for identical repeat runs of the function;
else \code{NULL}.}

\item{plot}{A logical indicating whether to generate a plot of the phylogenetic tree (\code{TRUE}) or not (\code{FALSE}, the default).}
}
\description{
Currently under development. Please use the regular snp.sim function to simulate genetic data.
}
\examples{
## Example ##

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
