% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{spaVS}
\alias{spaVS}
\title{Title}
\usage{
spaVS(data, fit_X, fit_y_given_X, family = "binomial")
}
\arguments{
\item{data}{}

\item{fit_X}{}

\item{fit_y_given_X}{}

\item{family}{}
}
\description{
Title
}
