% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{tower_gcm}
\alias{tower_gcm}
\title{GCM doubly robust variable selection function}
\usage{
tower_gcm(
  data,
  fit_X,
  fit_y_given_X,
  K = 5,
  alpha = 0.1,
  multiple_correction = "BH",
  seed = 1234
)
}
\arguments{
\item{data}{A list with an n x p data matrix X and n x 1 vector y}

\item{fit_X}{A function that inputs X and outputs a function
that inputs X, j, B and outputs B conditional resamples of
variable j.}

\item{fit_y_given_X}{A function that inputs X and y and family and
outputs a function that inputs X and outputs the fitted
E(y|X)}

\item{K}{The number of folds for crossfitting}

\item{alpha}{Desired FDP level.}

\item{multiple_correction}{How to adjust the p values for multiplicity.}

\item{seed}{Seed to set for data split and fits of Y|X and X}

\item{family}{A string specifying the exponential family, e.g. "gaussian", "bernoulli"}

\item{B}{The number of resamples}

\item{linear}{whether the fit_y_given_X is linear in X}
}
\value{
A list of p p-values, one for each variable
}
\description{
GCM doubly robust variable selection function
}
