% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{tree_prune}
\alias{tree_prune}
\title{Title}
\usage{
tree_prune(tree, labels)
}
\arguments{
\item{tree}{a phylogenetic tree}

\item{labels}{labels of the tree to keep}
}
\value{
a pruned tree which only keeps data points in labels
}
\description{
Title
}
