% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{treeWAS_no_subs_wrapper}
\alias{treeWAS_no_subs_wrapper}
\title{Title}
\usage{
treeWAS_no_subs_wrapper(
  data,
  method = c("terminal", "simultaneous", "subsequent"),
  n.snps.sim = 10000
)
}
\arguments{
\item{data}{}
}
\description{
Title
}
