% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{treeWAS_tree_wrapper}
\alias{treeWAS_tree_wrapper}
\title{Title}
\usage{
treeWAS_tree_wrapper(
  data,
  method = c("terminal", "simultaneous", "subsequent"),
  n.snps.sim = 10000
)
}
\arguments{
\item{data}{a list object that contains an X matrix, y vector, phylogenetic tree.
The rownames of X and names of y should align, and be a subset of tree$tip.label}

\item{method}{vector of scores to compute in treeWAS}

\item{n.snps.sim}{number of samples to use}
}
\value{
output from treeWAS
}
\description{
Title
}
