% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spa_helpers.R
\name{wcgf}
\alias{wcgf}
\title{\code{wcgf} is a function computing the cumulant generating function (CGF) of
distributions, multiplied by a weight function, from the GLM family}
\usage{
wcgf(s, P, W, fam)
}
\arguments{
\item{s}{The point where the CGF will be computed.}

\item{P}{A vector containing the parameter values of the family of distributions.}

\item{W}{A vector containing the weights.}

\item{fam}{The GLM family which includes the distribution whose CGF is being
evaluated (values can be \code{gaussian}, \code{binomial}, \code{poisson}, etc).}
}
\value{
CGF of the weighted distribution evaluated at \code{s}.
}
\description{
\code{wcgf} is a function computing the cumulant generating function (CGF) of
distributions, multiplied by a weight function, from the GLM family
}
\keyword{internal}
