% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{gmmat_binary}
\alias{gmmat_binary}
\title{Title}
\usage{
gmmat_binary(
  data,
  covar = NULL,
  phylo = FALSE,
  jaccard = FALSE,
  hamming = FALSE,
  normalized = FALSE,
  tmpdir = "gmmat_temp",
  output = "gmmat_out.txt"
)
}
\arguments{
\item{covar}{}

\item{y}{binary phenotype}

\item{X}{genetic features matrix}

\item{verbose}{}
}
\description{
Title
}
