% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{pyseer_wrapper}
\alias{pyseer_wrapper}
\title{Title}
\usage{
pyseer_wrapper(
  data,
  method,
  tmpdir = "pyseer_temp",
  output = "pyseer_out.txt",
  conda_bin,
  pyseer_env,
  max_dimensions = 10,
  phylo = FALSE,
  jaccard = FALSE,
  hamming = FALSE,
  normalized = FALSE,
  extra_args = NULL,
  keep_files = FALSE
)
}
\arguments{
\item{data}{}

\item{method}{}

\item{tmpdir}{}

\item{output}{}

\item{conda_bin}{}

\item{pyseer_env}{}

\item{extra_args}{}

\item{keep_files}{}
}
\description{
Title
}
