% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{susieK}
\alias{susieK}
\title{Title}
\usage{
susieK(
  data,
  L,
  K = NULL,
  phylo = TRUE,
  est_ssq = TRUE,
  ssq = NULL,
  ssq_range = c(0, 1),
  pi0 = NULL,
  est_sigmasq = TRUE,
  est_tausq = TRUE,
  sigmasq = 1,
  tausq = 0,
  method = "moments",
  sigmasq_range = NULL,
  tausq_range = NULL,
  PIP = NULL,
  mu = NULL,
  maxiter = 100,
  PIP_tol = 0.001,
  LD = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{a list containing a vector of outcomes y, data matrix X, and optionally a tree}

\item{L}{number of modeled causal effects}

\item{K}{a similarity matrix}

\item{phylo}{whether to use the given tree to construct the similarity matrix K}

\item{est_ssq}{estimate prior effect size variances s^2 using MLE}

\item{ssq}{length-L initialization s^2 for each effect Default: 0.2 for every effect}

\item{ssq_range}{lower and upper bounds for each s^2, if estimated}

\item{pi0}{length-p vector of prior causal probability for each SNP; must sum to 1 Default: 1/p for every SNP}

\item{est_sigmasq}{estimate variance sigma^2}

\item{est_tausq}{estimate both variances sigma^2 and tau^2}

\item{sigmasq}{initial value for sigma^2}

\item{tausq}{initial value for tau^2}

\item{method}{one of {'moments','MLE'} (sigma^2,tau^2) are estimated using method-of-moments or MLE}

\item{sigmasq_range}{lower and upper bounds for sigma^2, if estimated using MLE}

\item{tausq_range}{lower and upper bounds for tau^2}

\item{PIP}{p x L initializations of PIPs Default: 1/#SNPs for each SNP and effect}

\item{mu}{p x L initializations of mu; Default: 0 for each SNP and effect}

\item{maxiter}{maximum number of SuSiE iterations}

\item{PIP_tol}{convergence threshold for PIP difference between iterations}

\item{verbose}{}
}
\value{
List containing PIP -- p x L matrix of PIPs, individually for each effect mu -- p x L matrix of posterior means conditional on causal omega -- p x L matrix of posterior precisions conditional on causal lbf -- length-L array of log-Bayes-factor for each effect lbf_variable -- p x L matrix of per-variable log-Bayes-factors ssq -- length-L array of final effect size variances s^2 sigmasq -- final value of sigma^2 tausq -- final value of tau^2 alpha -- length-p array of posterior means of infinitesimal effects
}
\description{
Title
}
